+++ CHANGELOG.md
@@ -348,8 +348,15 @@ Change main.c to use OUTPUTTYPE instead of float
 - HOTFIX: Deeploy subdirectories installed when installing Deeploy with pip install
 - HOTFIX: Linking TEST_RECENT on MacOS
 
-
 ## Add RV32IMF Picolibc support for Siracusa platform
 
 ### Added
 - Adds RV32IMF Picolib to the toolchain
+
+## Add Generic Float DW Conv2D Kernel
+
+### Added
+- Generic float DW Conv2D kernel and bindings.
+- Bias handling and computation for regular and DW Conv2D.
+- Empty bias handling for generic regular and DW Conv2D.
+- Tests for Conv2D regular and DW, with and without bias (and included them in the CI pipeline).